Delete Maintained a map with integrated search functionality for universities and transition view to current candidates from those universities. Wrote native Javascript AJAX code for the map search functionality (currently responsible for migrating to AngularJS) Responsible for RESTful NodeJS services (using Express) which connected to MongoDB for performing document query and delivering it to the front end. Used google maps API to reverse geocode the address (obtained from MongoDB collection) and display that as a marker on the map Used async.js to run AJAX calls in series so that the google api limit is not reached Maintained the PassportJS for authentication of users Built the front end using CSS3 and HTML5 tagsREADME.md
Use nodeJS, googleMap and MongoDB to implement a University Search project Maintained a map with integrated search functionality for universities and transition view to current candidates from those universities. Wrote native Javascript AJAX code for the map search functionality (currently responsible for migrating to AngularJS) Responsible for RESTful NodeJS services (using Express) which connected to MongoDB for performing document query and delivering it to the front end. Used google maps API to reverse geocode the address (obtained from MongoDB collection) and display that as a marker on the map Used async.js to run AJAX calls in series so that the google api limit is not reached Maintained the PassportJS for authentication of users Built the front end using CSS3 and HTML5 tags